added a folder for compare components in the componenets folder
