+++ frontend/templates/compare/AboutUse.tsx
@@ -1,17 +1,13 @@
 import { ReactElement } from 'react';
 import styled from 'styled-components';
 import 'antd/dist/antd.css';
-import { Row, Col } from 'antd';
 import SvgRemoveImplant from '../../public/remove-implant.svg';
 import SvgRollCondom from '../../public/roll-condom.svg';
+import TwoColumns from '../../components/compare/TwoColumns';
+import Title from '../../components/compare/Title';
+import { Text } from '../../pages/compare';
 
 // Styling
-
-const Header = styled.h1`
-  padding-top: 1vh;
-  padding-bottom: 3vh;
-  margin: 0;
-`;
 
 const RemoveImplantImage = styled(SvgRemoveImplant)`
   display: block;
@@ -25,11 +21,6 @@ const RollCondomImage = styled(SvgRollCondom)`
   margin-bottom: 2vh;
   margin-left: auto;
   margin-right: auto;
-`;
-
-const Text = styled.p`
-  max-width: 400px;
-  margin: 0 auto;
 `;
 
 export interface AboutUseProps {
@@ -49,40 +40,16 @@ const AboutUse = ({
   usePatternUnits,
   usePatternUnitsRight,
 }: AboutUseProps): ReactElement => {
-  const Title = (title: string): ReactElement => {
-    return (
-      <Col span={24}>{<Header className="centerText">{title}</Header>}</Col>
-    );
-  };
-  const ColText = (
-    leftText: ReactElement,
-    rightText: ReactElement,
-  ): ReactElement => {
-    return (
-      <Row>
-        <Col span={9} offset={2}>
-          {leftText}
-        </Col>
-        <Col span={9} offset={2}>
-          {rightText}
-        </Col>
-      </Row>
-    );
-  };
   return (
     <div>
       {Title('How To Use')}
-      <Row>
-        <Col span={9} offset={2}>
-          <RollCondomImage></RollCondomImage>
-        </Col>
-        <Col span={9} offset={2}>
-          <RemoveImplantImage></RemoveImplantImage>
-        </Col>
-      </Row>
-      {ColText(<Text>{howItWorks}</Text>, <Text>{howItWorksRight}</Text>)}
+      {TwoColumns(
+        <RollCondomImage></RollCondomImage>,
+        <RemoveImplantImage></RemoveImplantImage>,
+      )}
+      {TwoColumns(<Text>{howItWorks}</Text>, <Text>{howItWorksRight}</Text>)}
       {Title('How often to use?')}
-      {ColText(
+      {TwoColumns(
         <Text className="centerText bold teal">
           Lasts up to {usePatternHighBound} {usePatternUnits}
         </Text>,
@@ -90,7 +57,7 @@ const AboutUse = ({
           Lasts up to {usePatternHighBoundRight} {usePatternUnitsRight}
         </Text>,
       )}
-      {ColText(
+      {TwoColumns(
         <Text>
           Placed on male tip of the penis, roll it down over the length of the
           erect penis.
